zorko-designer-105 Organize folders according to Docs - move logger and analytic to packages folder
