+++ src/features/chooseSpecFlow/actions.ts
@@ -1,5 +1,5 @@
 import {createAction} from '@reduxjs/toolkit';
-import {VegaLiteTopLevelUnitSpec} from '../../common/types';
+import {VegaLiteTopLevelUnitSpec} from '../../packages/coreTypes/types';
 
 export const chooseSpecFlowOpenSpecs = createAction('chooseSpecFlow/openSpecs');
