+++ src/features/dataSourceMetadata/asyncActions.ts
@@ -1,7 +1,11 @@
-import * as Api from '../../api';
+import * as Api from '../../packages/designerApi';
 import {DataSource} from 'vega-lite/build/src/data';
 import logger from 'logrock';
-import {dataSourceMetadataReadFailure, dataSourceMetadataReadRequest, dataSourceMetadataReadSuccess} from './actions';
+import {
+  dataSourceMetadataReadFailure,
+  dataSourceMetadataReadRequest,
+  dataSourceMetadataReadSuccess
+} from './actions';
 
 export const dataSourceMetadataRead = (dataSource: DataSource) => {
   return async dispatch => {
